UIK-42: Extract methods to external script
